pytorch-dataset/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
filelock==3.13.1
fsspec==2024.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
setuptools==70.0.0
sympy==1.13.1
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
typing_extensions==4.9.0
